requirements.txt
# ===== Core Framework =====
flask==3.1.0
flask-cors==5.0.1
flask-limiter==3.9.0
flask-sqlalchemy==3.1.1

# ===== Database =====
SQLAlchemy==2.0.36

# ===== Authentication =====
PyJWT==2.10.1

# ===== Web Scraping =====
requests==2.32.3
beautifulsoup4==4.12.3
lxml==5.3.0
fake-useragent==2.0.3

# ===== NLP & Sentiment Analysis =====
textblob==0.18.0
vaderSentiment==3.3.2
nltk==3.9.1

# ===== Machine Learning =====
scikit-learn==1.6.1
pandas==2.2.3
numpy==2.2.3
joblib==1.4.2

# ===== AI Summarization =====
google-generativeai==0.8.4

# ===== Frontend =====
streamlit==1.41.1
plotly==5.24.1

# ===== Utilities =====
python-dotenv==1.0.1
marshmallow==3.23.2

# ===== Testing =====
pytest==8.3.4
pytest-cov==6.0.0

# ===== Dev Tools =====
black==24.10.0
flake8==7.1.1
